requirements.txt
aiohttp==3.6.3
arrow==0.17.0
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
discord.py==1.5.1
google==3.0.0
idna==2.10
iniconfig==1.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
mysql==0.0.2
mysql-connector-python==8.0.22
mysqlclient==2.0.1
packaging==20.4
peewee==3.14.0
pip==20.2.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
requests==2.24.0
rope==0.18.0
schedule==0.6.0
setuptools==49.2.1
six==1.15.0
soupsieve==2.0.1
tenorpy==1.0.3
toml==0.10.2
urllib3==1.25.11
vk==2.0.2
wikipedia==1.4.0
Wikipedia-API==0.5.4
wrapt==1.12.1
yarl==1.5.1